♻️ refactor(config): move api key manager into config module
- remove src/asksage_proxy/api_key_manager.py and integrate logic into config.py
- add api key load balancing and selection methods to AskSageConfig
- centralize api key management for easier access and code maintainability

♻️ refactor(models): eliminate model registry api key manager

- remove ModelRegistry init and usage of separate api key manager
- rely on config.api_key for AskSageClient initialization

♻️ refactor(chat): use config api key for chat completions

- remove reference to model_registry.api_key_manager in chat endpoint
- simplify API key retrieval for requests
